lame encoder replacemant

+++ legacy-projects/BSE.CDDrives/BSE.CDDrives/NativeMethods.cs
@@ -152,7 +152,7 @@ namespace BSE.CDDrives
                     try
                     {
                         IntPtr buffer = handle.AddrOfPinnedObject();
-                        buffer = (IntPtr)(buffer.ToInt32() + (Index * Marshal.SizeOf(typeof(TRACK_DATA))));
+                        buffer = (IntPtr)(buffer.ToInt64() + (Index * Marshal.SizeOf(typeof(TRACK_DATA))));
                         res = (TRACK_DATA)Marshal.PtrToStructure(buffer, typeof(TRACK_DATA));
                     }
                     finally
